move auto_taper to separate file
